+++ src/Sulu/Bundle/ContactBundle/Entity/AccountRepository.php
@@ -10,15 +10,15 @@
 
 namespace Sulu\Bundle\ContactBundle\Entity;
 
-use Doctrine\ORM\EntityRepository;
 use Doctrine\ORM\NoResultException;
 use Doctrine\ORM\Query;
+use Gedmo\Tree\Entity\Repository\NestedTreeRepository;
 
 /**
  * Repository for the Codes, implementing some additional functions
  * for querying objects
  */
-class AccountRepository extends EntityRepository
+class AccountRepository extends NestedTreeRepository
 {
     /**
      * Searches for accounts with a specific contact
@@ -28,7 +28,12 @@ class AccountRepository extends EntityRepository
     public function findOneByContactId($contactId)
     {
         $qb = $this->createQueryBuilder('a')
-            ->join('a.accountContacts', 'accountContacts', 'WITH', 'accountContacts.idContacts = :contactId AND accountContacts.main = TRUE')
+            ->join(
+                'a.accountContacts',
+                'accountContacts',
+                'WITH',
+                'accountContacts.idContacts = :contactId AND accountContacts.main = TRUE'
+            )
             ->setParameter('contactId', $contactId);
         $query = $qb->getQuery();
 
@@ -61,6 +66,8 @@ class AccountRepository extends EntityRepository
     {
         try {
             $qb = $this->createQueryBuilder('account')
+                ->leftJoin('account.categories', 'categories')
+                ->leftJoin('categories.translations', 'categoryTranslations')
                 ->leftJoin('account.accountAddresses', 'accountAddresses')
                 ->leftJoin('accountAddresses.address', 'addresses')
                 ->leftJoin('addresses.country', 'country')
@@ -75,15 +82,13 @@ class AccountRepository extends EntityRepository
                 ->leftJoin('account.notes', 'notes')
                 ->leftJoin('account.faxes', 'faxes')
                 ->leftJoin('faxes.faxType', 'faxType')
-                ->leftJoin('account.accountCategory', 'accountCategory')
                 ->leftJoin('account.bankAccounts', 'bankAccounts')
                 ->leftJoin('account.tags', 'tags')
-                ->leftJoin('account.termsOfDelivery', 'termsOfDelivery')
-                ->leftJoin('account.termsOfPayment', 'termsOfPayment')
-                ->leftJoin('account.responsiblePerson', 'responsiblePerson')
                 ->leftJoin('account.mainContact', 'mainContact')
                 ->leftJoin('account.medias', 'medias')
                 ->addSelect('mainContact')
+                ->addSelect('categories')
+                ->addSelect('categoryTranslations')
                 ->addSelect('partial tags.{id, name}')
                 ->addSelect('bankAccounts')
                 ->addSelect('accountAddresses')
@@ -100,10 +105,6 @@ class AccountRepository extends EntityRepository
                 ->addSelect('faxes')
                 ->addSelect('faxType')
                 ->addSelect('notes')
-                ->addSelect('accountCategory')
-                ->addSelect('termsOfDelivery')
-                ->addSelect('termsOfPayment')
-                ->addSelect('responsiblePerson')
                 ->addSelect('medias')
                 ->where('account.id = :accountId');
 
@@ -144,6 +145,7 @@ class AccountRepository extends EntityRepository
             }
 
             $query = $qb->getQuery();
+
             return $query->getResult();
         } catch (NoResultException $ex) {
             return null;
@@ -206,8 +208,8 @@ class AccountRepository extends EntityRepository
                 ->leftJoin('emails.accounts', 'emailsAccounts')
                 ->leftJoin('account.notes', 'notes')
                 ->leftJoin('account.bankAccounts', 'bankAccounts')
-                ->leftJoin('account.accountContacts','accountContacts')
-                ->leftJoin('accountContacts.contact','contacts')
+                ->leftJoin('account.accountContacts', 'accountContacts')
+                ->leftJoin('accountContacts.contact', 'contacts')
                 ->leftJoin('account.mainContact', 'mainContact')
                 ->leftJoin('accountContacts.position', 'position')
                 ->addSelect('position')
@@ -262,8 +264,8 @@ class AccountRepository extends EntityRepository
         try {
             $qb = $this->createQueryBuilder('account')
                 ->leftJoin('account.children', 'children')
-                ->leftJoin('account.accountContacts','accountContacts')
-                ->leftJoin('accountContacts.contact','contacts')
+                ->leftJoin('account.accountContacts', 'accountContacts')
+                ->leftJoin('accountContacts.contact', 'contacts')
                 ->select('count(DISTINCT children.id) AS numChildren')
                 ->addSelect('count(DISTINCT contacts.id) AS numContacts')
                 ->where('account.id = :accountId');
@@ -288,8 +290,8 @@ class AccountRepository extends EntityRepository
         try {
             $qb = $this->createQueryBuilder('account')
                 ->leftJoin('account.children', 'children')
-                ->leftJoin('account.accountContacts','accountContacts')
-                ->leftJoin('accountContacts.contact','contacts')
+                ->leftJoin('account.accountContacts', 'accountContacts')
+                ->leftJoin('accountContacts.contact', 'contacts')
                 ->select('account')
                 ->addSelect('children')
                 ->addSelect('accountContacts')